feat: rename InputText Component
