+++ packages/components/src/components/atoms/form/InputTextWithoutError.test.tsx
@@ -1,0 +1,43 @@
+import React from 'react'
+
+import { renderWithProviders } from '../../../utils/tests'
+
+import InputTextWithoutError from './InputTextWithoutError'
+
+describe('InputText match snapshot', () => {
+  it('BadgeDisabledFalse', () => {
+    const { container } = renderWithProviders(
+      <InputTextWithoutError
+        id="input-text"
+        label="Label"
+        placeholder="Name"
+        required={false}
+        disabled={false}
+        helperText=""
+        helperIcon=""
+        color="primary"
+        endAdornment={null}
+        value="hello world"
+      />
+    )
+    expect(container).toMatchSnapshot()
+  })
+
+  it('BadgeDisabledFalseRequired', () => {
+    const { container } = renderWithProviders(
+      <InputTextWithoutError
+        id="input-text"
+        label="Label"
+        placeholder="Name"
+        required
+        disabled
+        helperText=""
+        helperIcon=""
+        color="primary"
+        endAdornment={null}
+        value="hello world"
+      />
+    )
+    expect(container).toMatchSnapshot()
+  })
+})
