+++ packages/components/src/components/index.ts
@@ -13,8 +13,8 @@ export { default as DoubleDatePicker } from './atoms/form/DoubleDatePicker'
 export { default as DropDownWithoutError } from './atoms/form/DropDownWithoutError'
 export { default as DropDown } from './atoms/form/DropDown'
 export { default as InfoTooltip } from './atoms/form/InfoTooltip'
+export { default as InputTextWithoutError } from './atoms/form/InputTextWithoutError'
 export { default as InputText } from './atoms/form/InputText'
-export { default as InputTextError } from './atoms/form/InputTextError'
 export { default as RadioGroup } from './atoms/form/RadioGroup'
 export { default as Range } from './atoms/form/Range'
 export { default as RangeError } from './atoms/form/RangeError'
